bertopic/requirements.txt
aiofiles==22.1.0
aiohttp==3.9.1
aiosignal==1.3.1
aiosqlite==0.18.0
alabaster==0.7.12
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.14.2
asttokens==2.0.5
atomicwrites==1.4.0
attrs==23.1.0
autopep8==1.6.0
Babel==2.11.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
bertopic==0.15.0
binaryornot==0.4.4
black==23.11.0
bleach==4.1.0
Bottleneck==1.3.5
Brotli==1.0.9
certifi==2023.11.17
cffi==1.16.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.2
contourpy==1.2.0
cookiecutter==2.5.0
cryptography==41.0.7
cycler==0.11.0
Cython==3.0.7
datasets==2.14.4
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.7
docstring-to-markdown==0.11
docutils==0.18.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.13.0
flake8==6.0.0
fonttools==4.25.0
frozenlist==1.4.1
fsspec==2023.10.0
googlepatentscraper==0.2
hdbscan==0.8.33
huggingface-hub==0.17.3
idna==3.4
imagesize==1.4.1
importlib-metadata==6.0.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.25.0
ipython==8.15.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isort==5.9.3
jaraco.classes==3.2.1
jedi==0.18.1
jellyfish==1.0.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.6
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter_server==2.10.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyter-ydoc==0.2.4
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.9
keybert==0.8.3
keyring==23.13.1
kiwisolver==1.4.4
lazy-object-proxy==1.6.0
llvmlite==0.41.0
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.4
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
munkres==1.1.4
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.6
networkx==3.2.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.58.1
numexpr==2.8.7
numpy==1.26.2
numpydoc==1.5.0
overrides==7.4.0
packaging==23.1
pandas==2.1.4
pandocfilters==1.5.0
paramiko==2.8.1
parso==0.8.3
pathspec==0.10.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.3.1
platformdirs==3.10.0
plotly==5.18.0
pluggy==1.0.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.36
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.3.0
pyflakes==3.0.1
Pygments==2.15.1
pylint==2.16.2
pylint-venv==2.3.0
pyls-spyder==0.4.0
PyNaCl==1.5.0
pynndescent==0.5.11
pyOpenSSL==23.2.0
pyparsing==3.0.9
PyQt5==5.15.10
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.2
python-slugify==5.0.2
pytoolconfig==1.2.6
pytz==2023.3.post1
pywin32==305.1
pywin32-ctypes==0.2.0
pywinpty==2.0.10
PyYAML==5.4.1
pyzmq==25.1.0
QDarkStyle==3.0.2
qstylizer==0.2.2
QtAwesome==1.2.2
qtconsole==5.4.2
QtPy==2.4.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rope==1.7.0
rpds-py==0.10.6
Rtree==1.0.1
safetensors==0.4.0
scikit-learn==1.2.2
scipy==1.11.4
seaborn==0.12.2
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.2.2
sip==6.7.12
six==1.16.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==5.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.4.3
spyder-kernels==2.4.4
stack-data==0.2.0
statsmodels==0.14.0
sympy==1.12
tenacity==8.2.3
terminado==0.17.1
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==2.2.0
three-merge==0.1.1
tinycss2==1.2.1
tokenizers==0.14.1
toml==0.10.2
tomlkit==0.11.1
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.7.1
transformers==4.34.1
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.4.0
umap-learn==0.5.5
Unidecode==1.2.0
urllib3==1.26.18
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
whatthepatch==1.0.2
wheel==0.41.2
widgetsnbextension==4.0.5
win-inet-pton==1.1.0
wrapt==1.14.1
xxhash==3.4.1
y-py==0.5.9
yapf==0.31.0
yarl==1.9.3
ypy-websocket==0.8.2
zipp==3.11.0
